+++ src/extension.ts
@@ -4,11 +4,11 @@ import * as vscode from "vscode";
 
 import { AppModel } from "./appModel";
 import { checkNewAnnouncement } from "./announcement/index";
-import { ErrorLogger } from "./VscodeExtensions";
+import { ErrorLogger, OutputLevel, OutputWindow } from "./VscodeExtensions";
 
 export async function activate(context: vscode.ExtensionContext): Promise<void> {
     try {
-        console.log('"live-sass-compiler" is now activated! Go and debug :P ');
+        OutputWindow.Show(OutputLevel.Trace, '"live-sass-compiler" is now activate');
 
         const appModel = new AppModel(context.workspaceState);
 
@@ -63,7 +63,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
                 }
             ),
             disposableOnDidSave = vscode.workspace.onDidSaveTextDocument(async () => {
-                // TODO: ADD - once autopefixer can stop caching browserslist
+                // TODO: ADD - once autoprefixer can stop caching browserslist
                 //await appModel.browserslistChecks();
                 await appModel.compileOnSave();
             });
@@ -80,6 +80,10 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
             disposableDebugFileList,
             appModel
         );
+
+        OutputWindow.Show(OutputLevel.Trace, "Live SASS commands ready", [
+            "Commands have been saved and are ready to be used",
+        ]);
     } catch (err) {
         await new ErrorLogger(context.workspaceState).LogIssueWithAlert(
             `Unhandled error with Live Sass Compiler. Error message: ${err.message}`,
@@ -91,5 +95,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
 }
 
 export function deactivate(): void {
-    // No action required
+    // No actual actions are required
+
+    OutputWindow.Show(OutputLevel.Trace, '"live-sass-compiler" deactivated');
 }